% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP_routines.R
\name{release_dataset_all_numeric}
\alias{release_dataset_all_numeric}
\title{Release an all numeric dataset.}
\usage{
release_dataset_all_numeric(
  data,
  sensitivity = 1,
  diffpriv_mechanism = diffpriv::DPMechGausian,
  diffpriv_pparams = diffpriv::DPParamsEps(epsilon = 5)
)
}
\arguments{
\item{diffpriv_mechanism}{(unused)}

\item{diffpriv_pparams}{(unused)}
}
\description{
Release an all numeric dataset.
}
